requirements.txt
git+ssh://git@github.com/rapptz/discord.py@async
pyyaml==3.11
